Fix path in links to exercise materials
The relative link from index(2?).html to the exercises is not under
'assets/' - this removes that from all the links.

There is also a reference to exercises/assets/cupcake.jpg in one of the
exercises, which does not exist.  This adds it a cupcake image from
wikipedia that has a creative commons zero license.